Make one models file
Manually resolve merge conflict
